Fixes https://github.com/tantaman/repliear-row-versioning/issues/6
Sequential scans put a read lock on the entire table, which when
combined with SERIALIZABLE means that two pulls can't happen at
the same time.

When Postgres has little data in a table it will tend to prefer
sequential scans as they are cheaper. But if there's so little
data that this matters an index scan is fine too.

I think that with SERIALIZABLE it makes sense to avoid sequential
scans by default.

For more information see:

https://stackoverflow.com/questions/42288808/why-does-postgresql-serializable-transaction-think-this-as-conflict
https://www.postgresql.org/docs/current/transaction-iso.html